+++ qa/rpc-tests/rawtransactions.py
@@ -10,6 +10,7 @@
 
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import *
+from binascii import unhexlify,hexlify
 
 # Create one-input, one-output, no-fee transaction:
 class RawTransactionsTest(BitcoinTestFramework):
@@ -56,6 +57,14 @@ class RawTransactionsTest(BitcoinTestFramework):
         rawtx   = self.nodes[2].createrawtransaction(inputs, outputs)
         rawtx   = self.nodes[2].signrawtransaction(rawtx)
 
+        # check missing input also using verifyrawtransactions
+        err   = self.nodes[2].verifyrawtransactions([rawtx['hex']])
+        assert(err is not None)
+        assert(err['index'] == 0)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('bad-txns-inputs-missingorspent'))
+
+        # check sendrawtransaction
         errorString = ""
         try:
             rawtx   = self.nodes[2].sendrawtransaction(rawtx['hex'])
@@ -63,6 +72,72 @@ class RawTransactionsTest(BitcoinTestFramework):
             errorString = e.error['message']
 
         assert("Missing inputs" in errorString)
+
+        #################################
+        # validaterawtransactions tests #
+        #################################
+        inputs  = []
+        outputs = { self.nodes[0].getnewaddress() : 4.998 }
+        rawtx   = self.nodes[2].createrawtransaction(inputs, outputs)
+
+        # check that unfunded transaction is invalid
+        err   = self.nodes[2].verifyrawtransactions([rawtx])
+        assert(err is not None)
+        assert(err['index'] == 0)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('bad-txns-vin-empty'))
+
+        rawtx   = self.nodes[2].fundrawtransaction(rawtx)
+        rawtx   = rawtx['hex']
+
+        # check that unsigned transaction is invalid
+        err   = self.nodes[2].verifyrawtransactions([rawtx])
+        assert(err is not None)
+        assert(err['index'] == 0)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('mandatory-script-verify-flag-failed (Operation not valid with the current stack size)'))
+
+        rawtx   = self.nodes[2].signrawtransaction(rawtx)
+        rawtx   = rawtx['hex']
+
+        # check that transaction is fully valid
+        err   = self.nodes[2].verifyrawtransactions([rawtx])
+        assert(err is None)
+
+        # check that duplicate transaction causes missing/spent inputs error
+        # in second transaction
+        err   = self.nodes[2].verifyrawtransactions([rawtx,rawtx])
+        assert(err is not None)
+        assert(err['index'] == 1)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('bad-txns-inputs-missingorspent'))
+
+        # corrupt the transaction output to invalidate the signature, will result in EVAL_FALSE
+        rawtx = bytearray(unhexlify(rawtx))
+        rawtx[-10] = rawtx[-10] ^ 0xff
+        rawtx = hexlify(rawtx)
+        err   = self.nodes[2].verifyrawtransactions([rawtx])
+        assert(err['index'] == 0)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('mandatory-script-verify-flag-failed (Script evaluated without error but finished with a false/empty top stack element)'))
+
+        # loose coinbase transaction should be rejected
+        rawtx = '01000000010000000000000000000000000000000000000000000000000000000000000000ffffffff4d04ffff001d0104455468652054696d65732030332f4a616e2f32303039204368616e63656c6c6f72206f6e206272696e6b206f66207365636f6e64206261696c6f757420666f722062616e6b73ffffffff0100f2052a01000000434104678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5fac00000000'
+        err   = self.nodes[2].verifyrawtransactions([rawtx])
+        assert(err['index'] == 0)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('coinbase'))
+
+        # passing no transactions at all is always successful, albeit boring
+        err   = self.nodes[2].verifyrawtransactions([])
+        assert(err is None)
+
+        # invalid option name
+        try:
+            err   = self.nodes[2].verifyrawtransactions([],{'check_transaction_color':False})
+            assert(false) # invalid option must cause exception
+        except JSONRPCException,e:
+            assert(e.error['code'] == -8) # RPC_INVALID_PARAMETER
 
         #########################
         # RAW TX MULTISIG TESTS #
@@ -132,12 +207,47 @@ class RawTransactionsTest(BitcoinTestFramework):
         
         rawTxSigned = self.nodes[2].signrawtransaction(rawTx, inputs)
         assert_equal(rawTxSigned['complete'], True) #node2 can sign the tx compl., own two of three keys
-        self.nodes[2].sendrawtransaction(rawTxSigned['hex'])
+
+        newTxId = self.nodes[2].sendrawtransaction(rawTxSigned['hex'])
+
+        # after syncing up mempools, check that verifyrawtransactions against
+        # mempool and against chain will now diverge when an output from the
+        # previously constructed transaction is used
+        self.sync_all()
+        inputs = [{ "txid" : newTxId, "vout" : 0 }]
+        outputs = { self.nodes[0].getnewaddress() : 2.19 }
+        rawTx2 = self.nodes[0].createrawtransaction(inputs, outputs)
+        rawTx2 = self.nodes[0].signrawtransaction(rawTx2)
+        rawTx2 = rawTx2['hex']
+        err   = self.nodes[0].verifyrawtransactions([rawTx2],{'include_mempool':True})
+        assert(err is None)
+        err   = self.nodes[0].verifyrawtransactions([rawTx2],{'include_mempool':False})
+        assert(err is not None)
+        assert(err['code'] == 16)
+        assert(err['reason'].startswith('bad-txns-inputs-missingorspent'))
+
+        #check ordering of dependent tx's in verifyrawtransactions
+        parent  = rawTxSigned['hex']
+        child   = rawTx2
+        err1     = self.nodes[2].verifyrawtransactions([parent, child], {'include_mempool':True})
+        err2     = self.nodes[2].verifyrawtransactions([parent, child], {'include_mempool':False})
+        assert(err1 is None and err2 is None)
+
+        # out-of-order succeeds with mempool, fails without
+        err1     = self.nodes[2].verifyrawtransactions([child, parent], {'include_mempool':True})
+        err2     = self.nodes[2].verifyrawtransactions([child, parent], {'include_mempool':False})
+        assert(err1 is None and err2 is not None)
+        assert(err2['reason'].startswith('bad-txns-inputs-missingorspent'))
+
+        # mine the transaction into a block and check end result
         rawTx = self.nodes[0].decoderawtransaction(rawTxSigned['hex'])
         self.sync_all()
         self.nodes[0].generate(1)
         self.sync_all()
         assert_equal(self.nodes[0].getbalance(), bal+Decimal('50.00000000')+Decimal('2.19000000')) #block reward + tx
+        # now that its ancester is in a block, the transaction should pass verification without mempool as well
+        err   = self.nodes[0].verifyrawtransactions([rawTx2],{'include_mempool':False})
+        assert(err is None)
 
 if __name__ == '__main__':
     RawTransactionsTest().main()